Move each Recognizer to a separate file & reorganize their code
