+++ packages/fetch-and-compile/lib/debug.ts
@@ -1,70 +1,14 @@
 import debugModule from "debug";
-const debug = debugModule("fetch-and-compile:recognizers");
+const debug = debugModule("fetch-and-compile:debug");
 import * as Codec from "@truffle/codec";
 import type {
   Recognizer,
   FailureType,
   FetchExternalErrors,
-  Instances,
-  SingleResult
+  Instances
 } from "./types";
 import type { WorkflowCompileResult } from "@truffle/compile-common";
 import type { SourceInfo } from "@truffle/source-fetcher";
-
-export class SingleRecognizer implements Recognizer {
-  private address: string;
-  private recognized: boolean = false;
-  private compileResult: WorkflowCompileResult;
-  private sourceInfo: SourceInfo;
-
-  constructor(address: string) {
-    this.address = address;
-  }
-
-  isAddressUnrecognized(address: string): boolean {
-    return !this.recognized || address !== this.address; //I guess?
-  }
-
-  getAnUnrecognizedAddress(): string | undefined {
-    return this.recognized ? undefined : this.address;
-  }
-
-  markUnrecognizable(address: string, reason?: FailureType): never {
-    //just throw...
-    if (reason) {
-      switch (reason) {
-        case "fetch":
-          throw new Error(`Error in fetching sources for ${address}`);
-        case "compile":
-          throw new Error(`Error in compiling sources for ${address}`);
-      }
-    } else {
-      throw new Error(`No verified sources found for ${address}`);
-    }
-  }
-
-  markBadFetcher(_fetcherName: string): void {
-    //do nothing
-  }
-
-  addCompiledInfo(
-    compileResult: WorkflowCompileResult,
-    sourceInfo: SourceInfo,
-    _address: string,
-    _fetcherName: string
-  ): void {
-    this.compileResult = compileResult;
-    this.sourceInfo = sourceInfo;
-  }
-
-  //not required by the interface!
-  getResult(): SingleResult {
-    return {
-      compileResult: this.compileResult,
-      sourceInfo: this.sourceInfo
-    };
-  }
-}
 
 export class DebugRecognizer implements Recognizer {
   private bugger: any; //sorry, we don't have a type for the debugger
@@ -77,7 +21,16 @@ export class DebugRecognizer implements Recognizer {
     this.bugger = bugger; //no clone, note!
   }
 
-  private getUnrecognizedAddresses(): string[] { //helper function
+  getErrors(): FetchExternalErrors {
+    return {
+      fetch: this.badFetchAddresses,
+      compile: this.badCompileAddresses,
+      fetchers: this.badFetchers
+    };
+  }
+
+  //helper method
+  private getUnrecognizedAddresses(): string[] {
     debug("getting unknown addresses");
     const instances: Instances = this.bugger.view(
       this.bugger.selectors.session.info.affectedInstances
@@ -87,6 +40,10 @@ export class DebugRecognizer implements Recognizer {
       .filter(([_, { contractName }]) => contractName === undefined)
       .map(([address, _]) => address);
   }
+
+  /*
+   * Interface methods follow
+   */
 
   isAddressUnrecognized(address: string): boolean {
     return this.getUnrecognizedAddresses().includes(address);
@@ -129,14 +86,5 @@ export class DebugRecognizer implements Recognizer {
       `externalFor(${address})Via(${fetcherName})`
     );
     await this.bugger.addExternalCompilations(shimmedCompilations);
-  }
-
-  //not required by the interface!
-  getErrors(): FetchExternalErrors {
-    return {
-      fetch: this.badFetchAddresses,
-      compile: this.badCompileAddresses,
-      fetchers: this.badFetchers
-    };
   }
 }
