+++ packages/fetch-and-compile/lib/index.ts
@@ -1,7 +1,8 @@
 import debugModule from "debug";
 const debug = debugModule("fetch-and-compile");
 import type Config from "@truffle/config";
-import * as Recognizers from "./recognizers";
+import { SingleRecognizer } from "./recognizer";
+import { DebugRecognizer } from "./debug";
 import { fetchAndCompileForRecognizer } from "./fetch";
 import type * as Types from "./types";
 
@@ -11,7 +12,7 @@ export async function fetchAndCompile(
   address: string,
   config: Config
 ): Promise<Types.SingleResult> {
-  const recognizer = new Recognizers.SingleRecognizer(address);
+  const recognizer = new SingleRecognizer(address);
   await fetchAndCompileForRecognizer(recognizer, config);
   return recognizer.getResult();
 }
@@ -22,7 +23,7 @@ export async function fetchAndCompileForDebugger(
   bugger: any, //sorry; this should be a debugger object
   config: Config
 ): Promise<Types.FetchExternalErrors> {
-  const recognizer = new Recognizers.DebugRecognizer(bugger);
+  const recognizer = new DebugRecognizer(bugger);
   await fetchAndCompileForRecognizer(recognizer, config);
   return recognizer.getErrors();
 }
